M2-Challenge-Hoelscher-Rachel/src/main/java/com/company/M2ChallengeHoelscherRachel/models/Month.java
package com.company.M2ChallengeHoelscherRachel.models;

import java.util.Objects;

public class Month {
    private Integer number;
    private String name;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public Integer getNumber() {
        return number;
    }

    public void setNumber(Integer number) {
        this.number = number;
    }

    public Month(Integer number, String name) {
        this.number = number;
        this.name = name;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Month month = (Month) o;
        return Objects.equals(name, month.name) && Objects.equals(number, month.number);
    }

    @Override
    public int hashCode() {
        return Objects.hash(name, number);
    }

    @Override
    public String toString() {
        return "Month{" +
                "name='" + name + '\'' +
                ", number=" + number +
                '}';
    }
}
